Change all ember-data dates to strings
I just want plain old dumb dates with no timezones or frills, but it's
too smart for that. I'm not writing something to turn Y-M-D format into
a Date if ember-data folks couldn't be bothered to do the same.

This may become painful if we start doing lots of date operations on the
frontend, but I don't anticipate much of that. In that case I'll revert
this and instead augment the date serializer/transform to be that little
bit more liberal. (I think it would still need a shim for the native
date picker, which outputs strings anyhow.)