Save geolocation with each transaction if allowed
The intent is to keep track of the textual location ("Whole Foods") and
the coordinate (41.9086926,-87.6529275) for each transaction. We'll then
be able to complete/suggest a location based on previous transactions,
eg. if there's a nearby Whole Foods or Target that the user has visited
before, we can figure that out.

Privacy aside, this will need a way to opt out per-transaction, since
transactions might not be recorded at point of sale. Obviously we'll
want a global opt-in for privacy, but the browser is good at that. This
feature sends nothing if geolocation is denied.