Add per-transaction coordinate toggle
Upon creation, the user can opt out of geolocation with the icon.

Disabling coordinate works both before and after the coordinate is
gathered, so there's no race condition where the coordinate is resolved
just as the transaction is submitted.

After creation, this icon is just an indicator for whether a coordinate
is on record.